requirements.txt
apscheduler
python-dateutil
requests
paho-mqtt